+++ Android/.idea/misc.xml
@@ -154,6 +154,7 @@
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.16" />
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.16" />
         <entry key="app/src/main/res/xml/widget_provider.xml" value="0.21467391304347827" />
+        <entry key="cardStackView/src/main/res/layout/overlay.xml" value="0.15081521739130435" />
         <entry key="shared/android/src/main/res/drawable/background_color_sound.xml" value="0.1" />
         <entry key="shared/android/src/main/res/drawable/drawable_search_word.xml" value="0.1" />
         <entry key="shared/android/src/main/res/drawable/ic_close.xml" value="0.1" />
