+++ Android/cardstackview/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="kr.co.dgsw.cardstackview"/>
+<manifest package="co.kr.cardstackview"/>
