+++ Android/cardstackview/src/main/java/co/kr/cardstackview/Direction.java
@@ -1,4 +1,4 @@
-package kr.co.dgsw.cardstackview;
+package co.kr.cardstackview;
 
 import java.util.Arrays;
 import java.util.List;
