+++ Android/cardstackview/src/main/java/co/kr/cardstackview/StackFrom.java
@@ -1,4 +1,4 @@
-package kr.co.dgsw.cardstackview;
+package co.kr.cardstackview;
 
 public enum StackFrom {
     None,
