+++ Android/cardstackview/src/main/java/co/kr/cardstackview/internal/AnimationSetting.java
@@ -1,8 +1,8 @@
-package kr.co.dgsw.cardstackview.internal;
+package co.kr.cardstackview.internal;
 
 import android.view.animation.Interpolator;
 
-import kr.co.dgsw.cardstackview.Direction;
+import co.kr.cardstackview.Direction;
 
 public interface AnimationSetting {
     Direction getDirection();
